Add inih and merge-credentials-file options to pg_dumpall
Due to limitations in CloudSQL we have the need to do effectively
a pg_dumpall but with the ability to not perform ALTER ROLE and to
merge passwords in to the statements. The alternative is to do this
at runtime before passing to the restoration tool (psql), but it's
nicer to have it at the source. This is pretty simple functionality.
This does vendor in the inih library. No idea if problematic.